pyproject.toml
[tool.poetry]
name = "web-blog"
version = "0.1.0"
description = ""
authors = ["marcosalves.dev <marcosalves.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2"
django-crispy-forms = "^2.4"
crispy-bootstrap5 = "^2025.4"
pillow = "^11.2.1"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^5.1.0"
django-browser-reload = "^1.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
